repo/backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
alembic
psycopg2-binary
python-multipart
passlib[bcrypt]
python-jose[cryptography]
bcrypt==4.0.1
passlib==1.7.4
pytest
httpx